+++ serial_port/serial_port_actions_direct.cpp
@@ -35,7 +35,7 @@ bool SerialPortActionsDirect::is_serial_port_open()
     return serial->isOpen();
 }
 
-bool SerialPortActionsDirect::set_kline_timings(unsigned long parameter, int value)
+bool SerialPortActionsDirect::set_kline_timings(uint32_t parameter, int value)
 {
     _P1_MAX = value;
     return STATUS_SUCCESS;
