+++ serial_port/serial_port_actions_direct.h
@@ -119,7 +119,7 @@ public:
 #define SERIAL_P4_MAX   0x07 // J2534 says this may not be changed
 
     uint8_t _P1_MAX = 10;
-    bool     set_kline_timings(unsigned long parameter, int value);
+    bool set_kline_timings(uint32_t parameter, int value);
 
     bool is_serial_port_open();
     int change_port_speed(QString portSpeed);
